Remove usage of the "wxc_custom.inc" file.
There is now a new file: "custom.inc" that can allow the user to make
site specific overrides, keep sql passwords out of the main ini file,
etc...
This is completely optional and not required. The file may even be
safely deleted if wanted.

Also some typo fixing in the README file.