get-map-info, map_display and map_functions updated.
Added "ini_set('memory_limit', '128M') to get-map-info and map_display.
If you have a large network, you may need to increase the amount of
memory needed. 128M should be more than enough in most cases, unless you
have a "Mega Mesh" like N2MH does. :)

Added "last_seen" info to the station popups (map_functions.inc)

Also added (hopefully) a catch in get-map-info to find a "stalled" node
polling run.
Sometimes the node polling portion of the script never completes and it
never gets a chance to unset the "currently_running" bit in the
database.
There is now a check so that if the "currently_running" bit is 1 and the
time the script last polled the node is greater than 3 * the
"node_polling_interval" (from the ini file), then the script will just
go out and poll the nodes as normal.
If it completes, it will unset the "currently_running" bit and will
update the timestamp.

These changes have all been suggested by N2MH. Thanks Mark! I finally
got to it!