Setting up for the abilty to tell if the client has internet access or not.
There is a way to have javascript give us some indication if the clients
browser has access to the greater internet or not. This would allow the
map webpage to either pull it's local resources or to use resources from
the internet.
For 1 you can get better Map layers if the client has internet access,
and why pull data over the mesh if we do not have to?