Changed to use default ini and other user editable files.
There is no longer a user-files.tar.gz file!

The user must remove the "-default" from the file extension, save the
file and then make their own changes. I have seen many other programs
use this and it seems to make sense.
The README has been updated to reflect this change.
Other ideas are appreciated still! :)

With the exception of user-settings.ini, the code will check for the
existance of the non "-default" files and use those instead of the
default ones.
Also if the user does not create a user-settings.ini file of their own
an error will be thrown telling them such.

Fixed the Popup for "total number of link" in the attributon bar.

Some new values were added to the meshmap.css file in order to make it
easier for the user to make customizations to the page if they want.
More of these will probably be added in the future.