Bug Fix: get-map-info.php, map-display.php
Due to commit before this I also noticed non-nodes were being put into
the database.
There were over a 100 in my database alone.
These are most likely non-node devices and will never be able to be
mapped.
Initially, I opted for a "save and recheck" style of polling, in the
hopes that something was just missed and it would get picked up later.
This is not the case and over time I think can see what is happening.

added a check in get-map-info.php to try and skip these devices if we
reach one, and due to that change, the numbers of shown and not show
node and links at the bottom right of the map is more accuate now (I
hope).

73