Enhancement: map_display.php
The check for internet access was using a POST variable to tell the page
if the client had access to the internet or not (it just tries to get
favicon.ico from google).
The problem with POST was that every time you refreshed the page, the
POST variable had to be resent, usually resulting in a dialog box from
the browser ("Do you want to resend this info?")

This drove me crazy, I could not stand that dialog box.

The check for internet access has been changed to use a cookie.
The cookie contains 1 item and expires after 30 minutes.
So after 30 minutes, the page will recheck for internet access and act
accordingly.

There is also a small change to the check for user-settings.ini.

The README.md/html files were slightly updated as well.