BugFix: More 5GHz channels, 900MHz nodes now found by 'board_id'
Added 'non-US' 5GHz channel numbers.

Device band was determined by channel, 900MHz and 2.4GHz have some of
the same channel number and the script had a hard time telling which was
which.
This has (hopefully) been fixed now. Since there are only 3 AREDN
compatible 900MHz devices, it was easier to just find those certain
board_id's. This also allowed channel #4 to be added back into the list
for 2.4GHz. (yes it was a hack but mostly worked :) )

Additionally, I've been trying to optimize the map_function.inc file.
A couple of the functions in that file consume 90%+ of the time it takes
to render the webpage. (load_LinkedTO() is a beast!)
On the cloud server this can sometimes completely stall the script(s)
and fail to load the page at all.