Bugfixes/Changes
Removed an errant "echo" that would cause output in "silent" mode.

When the polling script parsed through the interfaces looking for the
wlan MAC address and ip (and the LAN ip) it always had issues with the
different versions of the json file on different firmwares.
I think this has finally been solved.
The polling script should pull the proper info from even a 3.15.1.0 node
now and if those nodes have location info they should now appear on the
map too.
The admins of those devices should be strongly reminded to update!!

I mistakenly removed the red and green dot from the status area of the
admin page. They are restored now.

Parallel node polling will work now, but it is not recommended on large
networks, as there is still no control over how many scripts are
spawned. Small networks of should be just fine.
(this has been tested many times on a network of 20-30 (sometimes 100's)
with no issues)
Use at your own risk though.

New warnings in "Test Mode" to alert when a nodes location has been
"fixed" or not. It's very noticable now.

Some general clean up/typo fixing and removal of things that were long
due for removal.

fixed some file attributes.

Added some things to be used in the near future but are not done yet.

73