New Polling script! It will better catch all sysinfo.json values. It will also be easier to add more values later.
The new script does not yet find the LAN IP Address, with all the new devices there are several interfaces that could be the LAN interface.
LAN IP info isn't really needed for mapping anyways.

Added new stable firmware to the default user-settings.ini file, please update your file.
Added a new setting in the ini file "errorsInCron" that will have the script send errors when it is running from cron.
The system running the map script will have to have a local MTA (at least) setup, the messages will be received in the local mail of the user the script is running as (this is default behavior for cron).

Fixes #19
Fixes #12 (hopefully)
Fixes #10 (hopefully)

Enjoy!