enhancement/bugfix: Search for and zoom to node
Added search for and zoom to node button

needed to move the layer control button out of the way though.
while working on that found a way to disable the event that caused the layer control to expand when hovered over.
also forced the layer control popup  to *always* have a scroll bar, which should fix an issue with _some_ browsers...

Fixes #11